# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/alignment": {
            "get": {
                "description": "Returns byte alignment data from the service layer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memory"
                ],
                "summary": "Get byte alignment information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StructSizes"
                        }
                    }
                }
            }
        },
        "/memory": {
            "get": {
                "description": "Returns memory usage and GC statistics. Optional query parameter \"goroutine\" determines safe (GC-tracked) or unsafe (non-GC) allocation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memory"
                ],
                "summary": "Get memory usage",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Set to false to use unsafe memory allocation outside GC",
                        "name": "goroutine",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MemStats"
                        }
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "description": "Returns all todos with debug info",
                "summary": "Get all todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Debug"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new todo item. Request body must include the todo details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new todo",
                "parameters": [
                    {
                        "description": "Todo object",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Debug"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "Returns a single todo item by its ID. Optional query parameter 'goroutine' runs in a goroutine if true.",
                "tags": [
                    "Gouroutine example"
                ],
                "summary": "Get todo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Use goroutine",
                        "name": "goroutine",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Debug"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/unsafe": {
            "get": {
                "description": "Returns an integer that has been manipulated in an unsafe way",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "memory"
                ],
                "summary": "Byte manipulation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Debug": {
            "type": "object",
            "properties": {
                "data": {},
                "duration_ns": {
                    "type": "integer"
                }
            }
        },
        "models.MemStats": {
            "type": "object",
            "properties": {
                "after-alloc": {
                    "type": "string"
                },
                "after-gc": {
                    "type": "string"
                },
                "before-alloc": {
                    "type": "string"
                }
            }
        },
        "models.StructSizes": {
            "type": "object",
            "properties": {
                "normal-size": {
                    "type": "integer"
                },
                "normalsalignment": {
                    "type": "integer"
                },
                "packed-alignment": {
                    "type": "integer"
                },
                "packed-size": {
                    "type": "integer"
                }
            }
        },
        "models.Todo": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}